feat: updated pom.xml, init.sql, application.properties

+++ library-api/src/main/resources/init.sql
@@ -1,8 +1,9 @@
 create table book(
-    id bigint primary key,
+    id serial primary key,
     isbn varchar not null unique ,
     title varchar not null ,
     genre varchar not null ,
     description varchar not null ,
     author varchar not null
 )
+
